pyproject.toml
[project]
name = "news-aggregator"
version = "0.1.0"
description = "News Aggregator with AI Summarization"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "feedparser>=6.0.10",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.24",
    "alembic>=1.16.2",
    "detect-secrets>=1.5.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]
prod = ["fastapi", "uvicorn", "gunicorn"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
    "types-requests>=2.32.4.20250611",
]
